+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **v2.8.2**:
 - improve performance when loading item information like `Connections`, `Jobs`, `Schedules`, etc.
+- fixed issue when cancelling Spark and API Notebook cells
 
 **v2.8.1**:
 - fixed issue with `%run` where it was always resolved in the workspace of the lakehouse instead of relative to the current notebook